Fix bug in wind raster stack; remove any spatial data for Antarctica; update documentation
The wind data contain data for Antarctica, this is contrary to what the
README.txt file states. Any spatial data outputs are now cropped to the
same extent, of ymin = -60, ymax = 85, xmin = -180, xmax = 180.

This is now documented in help file for `create_CRU_stack`.

Documentation files are also update with a note explicitly stating that
elevation is converted to metres from kilometres in both
`create_CRU_df` and `create_CRU_stack`